#9 bomb position fixed to map
