+++ src/bomb/index.js
@@ -3,4 +3,5 @@ export { default as drawBomb } from './drawBomb';
 export { default as drawBombWithFuse } from './drawBombWithFuse';
 export { default as updateBomb } from './updateBomb';
 export { default as drawShrapnel } from './drawShrapnel';
+export { default as createShrapnel } from './createShrapnel';
 export * from './constants';
